move quell-client helper functions into helpers folder for better sturcturing
