+++ quell-client/Quellify.js
@@ -1,10 +1,10 @@
 const { parse } = require('graphql/language/parser');
-const parseAST = require('./parseAST');
-const normalizeForCache = require('./normalizeForCache');
-const buildArray = require('./buildArray');
-const createQueryObj = require('./createQueryObj');
-const createQueryStr = require('./createQueryStr');
-const joinResponses = require('./joinResponses');
+const parseAST = require('./helpers/parseAST');
+const normalizeForCache = require('./helpers/normalizeForCache');
+const buildArray = require('./helpers/buildArray');
+const createQueryObj = require('./helpers/createQueryObj');
+const createQueryStr = require('./helpers/createQueryStr');
+const joinResponses = require('./helpers/joinResponses');
 
 // NOTE:
 // Map: Query to Object Types map - Get from server or user provided (check introspection)
@@ -63,7 +63,7 @@ async function Quellify(endPoint, query, map, fieldsMap) {
     // If found data in cache:
     let mergedResponse;
     const queryObject = createQueryObj(proto); // Create query object from only false proto fields
-    console.log('queryObject in Quellify ===> ', queryObject)
+    console.log('queryObject in Quellify ===> ', queryObject);
     const queryName = Object.keys(proto)[0];
 
     // Partial data in cache:  (i.e. keys in queryObject will exist)
