+++ quell-client/helpers/buildArray.js
