+++ quell-client/helpers/createQueryObj.js
