+++ quell-client/helpers/createQueryStr.js
