+++ quell-client/helpers/joinResponses.js
