+++ quell-client/helpers/normalizeForCache.js
